代码微调5

+++ membership/membership-entity/src/main/java/com/tany/membership/entity/SysRole.java
@@ -1,9 +1,10 @@
 package com.tany.membership.entity;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import java.util.Date;
 import com.baomidou.mybatisplus.annotation.TableId;
+
 import java.io.Serializable;
+import java.util.Date;
 
 /**
  * <p>
@@ -18,7 +19,7 @@ public class SysRole implements Serializable {
     private static final long serialVersionUID=1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String code;
 
@@ -40,11 +41,11 @@ public class SysRole implements Serializable {
     private Date deletedDate;
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
